failed experiment: generating arbitrary well-typed terms via quickcheck
my suspicion from the start was that getting matching heads and bodies for the app case will be hard.
this is confirmed by one of the papers:
> decoupling the generation of types and expressions significantly biases generation towards functions that do not use their arguments.
perhaps I also made a silly circular dependency mistake in the generator on top of that.

some references (that I have not read, this impl was naive):
https://lemonidas.github.io/pdf/NotUseless.pdf
https://news.ycombinator.com/item?id=38038175
https://stackoverflow.com/q/57985003/8858828
https://github.com/Gabriella439/Haskell-Morte-Library/blob/3c61df86985a7ccf97fe0765deac73f06b12c476/test/ClosedWellTyped.hs#L38
https://github.com/bquiring/well-typed-term-generator
http://users.eecs.northwestern.edu/~baf111/random-judgments/random-judgments-esop15.pdf